+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 * Development task to rebuild the Algolia search index
 * Guided branch selection in select development task
 * ArcGIS scene layer and web map distribution options
+* Relationships between items presenting maps and their layers
 * Basic print styles
 
 ### Fixed